requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
praw==7.8.1
psycopg2-binary==2.9.9

# Telegram bot framework
python-telegram-bot==21.4



